src/rospatent_scraper/requirements.txt
fastapi
pydantic-settings
gunicorn
uvicorn
aiohttp
asyncpg
openpyxl
aioredis==1.3.1